refactor: 페이지 분리를 위한 React Router 적용
