+++ package.json
@@ -10,7 +10,8 @@
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.3.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.13.14",
