+++ src/App.js
@@ -1,25 +1,23 @@
 import React, { Component } from 'react';
-import EnterWinning from './pages/EnterWinning';
+import EnterWinning from './pages/EnterWinning/EnterWinning';
 import Main from './pages/Main/Main';
-import Result from './pages/Result';
-import { CURRENT_PAGE } from './constants';
+import Result from './pages/Result/Result';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 class App extends Component {
   constructor() {
     super();
-
-    this.state = {
-      currentPage: 'Main',
-    };
   }
 
   render() {
     return (
       <>
         <h1>행운의 로또</h1>
-        {this.state.currentPage === CURRENT_PAGE.MAIN && <Main />}
-        {this.state.currentPage === CURRENT_PAGE.ENTER_WINNING && <EnterWinning />}
-        {this.state.currentPage === CURRENT_PAGE.RESULT && <Result />}
+        <Router>
+          <Route exact path="/" component={Main} />
+          <Route exact path="/enter-winning" component={EnterWinning} />
+          <Route exact path="/result" component={Result} />
+        </Router>
       </>
     );
   }
